server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "aws-opentelemetry-distro>=0.10.0",
    "bedrock-agentcore>=1.0.3",
    "boto3",
    "mcp>=1.19.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]
